chore: reorganise game dev code
