+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "meta-gym-land",
-  "version": "0.0.1.5.22",
+  "version": "0.0.1.5.23",
   "dependencies": {
     "@ant-design/icons": "^4.7.0",
     "@ion-phaser/react": "^1.3.0",
