+++ src/components/Play/GymRoom.jsx
@@ -2,13 +2,13 @@ import React, { useState, useEffect, useContext } from "react";
 import { useMoralis } from "react-moralis";
 import Phaser from "phaser";
 import { IonPhaser } from "@ion-phaser/react";
-import { GymRoomScene } from "./GymRoomScene";
-import { SpaceStretchScene } from "./SpaceStretchScene";
-import { FlyFitScene } from "./FlyFitScene";
-import { ChartSquats } from "./ChartSquats";
-import { BootScene } from "./BootScene";
-import { MatrixScene } from "./MatrixScene";
-import { RushScene } from "./RushScene";
+import { GymRoomScene } from "../../games/gym-room/GymRoomScene";
+import { SpaceStretchScene } from "../../games/minigames/SpaceStretchScene";
+import { FlyFitScene } from "../../games/minigames/FlyFitScene";
+import { ChartSquats } from "../../games/minigames/ChartSquats";
+import { BootScene } from "../../games/gym-room-boot/BootScene";
+import { MatrixScene } from "../../games/minigames/MatrixScene";
+import { RushScene } from "../../games/minigames/RushScene";
 import { MiniGameCtx } from "index";
 import PoseDetWebcam from "components/Webcam/PoseDetWebcam";
 import SideMenu from "./GymRoomSideMenu";
