+++ src/components/Play/PlayPage.jsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import { useParams } from "react-router";
 import { AvatarCtx } from "index";
 import { Redirect } from "react-router";
-import GymRoom from "./games/GymRoom";
+import GymRoom from "./GymRoom";
 
 const PlayPage = () => {
   const [avatar] = useContext(AvatarCtx);
