+++ src/components/Play/images/move_right.png
