+++ src/components/Play/images/turn_right.png
