+++ src/games/gym-room-boot/assets/index.js
