+++ src/games/mechanics/moves-speed-caluclator.spec.js
