+++ src/games/minigames/RushSceneOld.js
@@ -1,13 +1,13 @@
 import Phaser from "phaser";
-import { getGameWidth, getGameHeight } from "./helpers";
-import { Player, RectObstacle } from "./objects";
-import { PLAYER_KEY, PLAYER_SCALE, GYM_ROOM_SCENE, RUSH } from "./shared";
+import { getGameWidth, getGameHeight } from "../helpers";
+import { Player, RectObstacle } from "../objects";
+import { PLAYER_KEY, PLAYER_SCALE, GYM_ROOM_SCENE, RUSH } from "../shared";
 import { createTextBox } from "./utils/text";
-import { mainBgColorNum, highlightTextColorNum } from "../../../GlobalStyles";
-import { EarnableScene } from "./EarnableScene";
-import * as gstate from "../../gpose/state";
-import * as gpose from "../../gpose/pose";
-import { RUSH_BG } from "./assets";
+import { mainBgColorNum, highlightTextColorNum } from "../../GlobalStyles";
+import { EarnableScene } from "./base-scenes/EarnableScene";
+import * as gstate from "../../components/gpose/state";
+import * as gpose from "../../components/gpose/pose";
+import { RUSH_BG } from "../gym-room-boot/assets";
 
 const SceneConfig = {
   active: false,
