+++ src/games/objects/index.js
