+++ src/games/objects/rect-obstacle.js
