+++ src/games/shared.js
