+++ src/games/utils/Globals.js
