+++ src/games/utils/collision_debugger.js
