+++ src/games/utils/text.js
